+++ backend/src/main/java/project/backend/service/impl/AppointmentServiceImpl.java
@@ -61,7 +61,7 @@ public class AppointmentServiceImpl implements AppointmentService {
     }
 
     @Override
-    public List<Appointment> getAppointmentByPatient(Long patientUtente) {
+    public List<Appointment> getAppointmentByPatient(int patientUtente) {
         return appointmentRepo.findByPatientNumUtente(patientUtente);
     }
